+++ src/api/controllers/auth.controller.ts
@@ -4,8 +4,8 @@ const httpStatus = require('http-status');
 import { User } from 'api/models';
 const RefreshToken = require('../models/refreshToken.model');
 const moment = require('moment-timezone');
-import { apiJson, slackWebhook } from 'api/utils/Utils';
-const { jwtExpirationInterval } = require('../../config/vars');
+import { apiJson, slackWebhook, sendEmail } from 'api/utils/Utils';
+const { JWT_EXPIRATION_MINUTES } = require('../../config/vars');
 
 /**
  * Returns a formated object with tokens
@@ -14,7 +14,7 @@ const { jwtExpirationInterval } = require('../../config/vars');
 function generateTokenResponse(user: any, accessToken: string) {
   const tokenType = 'Bearer';
   const refreshToken = RefreshToken.generate(user).token;
-  const expiresIn = moment().add(jwtExpirationInterval, 'minutes');
+  const expiresIn = moment().add(JWT_EXPIRATION_MINUTES, 'minutes');
   return {
     tokenType,
     accessToken,
@@ -35,6 +35,11 @@ exports.register = async (req: Request, res: Response, next: NextFunction) => {
     res.status(httpStatus.CREATED);
     const data = { token, user: userTransformed };
     // slackWebhook(`New User: ${user.email}`) // notify when new user registered
+    // sendEmail({
+    //   to: userTransformed.email, // for testing: only use authorized recipients in Mailgun Account Settings.
+    //   subject: 'Welcome!',
+    //   html: 'Welcome!'
+    // })
     return apiJson({ req, res, data });
   } catch (error) {
     return next(User.checkDuplicateEmail(error));